Queues for multiple thread communication

+++ thumbnail_maker.py
@@ -4,7 +4,9 @@ import os
 import logging
 from urllib.parse import urlparse
 from urllib.request import urlretrieve
-import threading
+from threading import Thread
+from queue import Queue
+
 
 import PIL
 from PIL import Image
@@ -21,52 +23,43 @@ class ThumbnailMakerService(object):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming' # Directory where image files are loaded from
         self.output_dir = self.home_dir + os.path.sep + 'outgoing' # Directory where resized images are placed
-        self.downloaded_bytes = 0
-        self.dl_lock = threading.Lock() # Locks threading resources to avoid threading interference
-        # Also referred to as Thread Synchronization
-        max_concurrent_dl = 4 # A limitation on the amount of concurrent downloads that can happen at one time
-        self.dl_sem = threading.Semaphore(max_concurrent_dl) # Semaphore threading for concurrent downloading limitation
-        
-    def download_image(self, url):
-        # download each image and save to the input dir 
-        self.dl_sem.acquire()
-        # Exception handling to prevent infinite locking of resources
-        try:
-            logging.info('Downloading image at URL ' + url)
-            img_filename = urlparse(url).path.split('/')[-1]
-            urlretrieve(url, self.input_dir + os.path.sep + img_filename)
-            # Counting the amount of bytes of downloaded images
-            dest_path = self.input_dir + os.path.sep + img_filename
-            img_size = os.path.getsize(dest_path)
-            with self.dl_lock: # with statement guarantees that an exception will not keep resources locked
-                self.downloaded_bytes += img_size
-            logging.info('Image [{} bytes] saved to '.format(img_size, dest_path))
-        finally:
-            self.dl_sem.release()
+        self.img_queue = Queue()
+        self.dl_queue = Queue()
+
+    def download_image(self):
+        while not self.dl_queue.empty():
+            '''
+            Avoids situation where 1 object left in queue and one thread
+            reads a valid value while another thread acceses and dequeues the object,
+            but the other thread proceeds to read
+            '''
+            try:
+                url = self.dl_queue.get(block = False)
+                # download each image and save to the input dir 
+                img_filename = urlparse(url).path.split('/')[-1]
+                urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+                self.img_queue.put(img_filename)
+
+                self.dl_queue.task_done()
+            except Queue.Empty:
+                logging.info('Queue Empty')
 
     def download_images(self, img_url_list):
         # validate inputs
-        if not img_url_list:
-            return
-        os.makedirs(self.input_dir, exist_ok=True)
-        
         logging.info("beginning image downloads")
 
         start = time.perf_counter()
-        threads = []
         for url in img_url_list:
-            t = threading.Thread(target = self.download_image, args = (url,))
-            t.start()
-            threads.append(t)
-        map(lambda t: t.join(), threads)
+            # download each image and save to the input dir 
+            img_filename = urlparse(url).path.split('/')[-1]
+            urlretrieve(url, self.input_dir + os.path.sep + img_filename)
+            self.img_queue.put(img_filename)
         end = time.perf_counter()
-
+        self.img_queue.put(None)
         logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
+	
 
     def perform_resizing(self):
-        # validate inputs
-        if not os.listdir(self.input_dir):
-            return
         os.makedirs(self.output_dir, exist_ok=True)
 
         logging.info("beginning image resizing")
@@ -74,32 +67,59 @@ class ThumbnailMakerService(object):
         num_images = len(os.listdir(self.input_dir))
 
         start = time.perf_counter()
-        for filename in os.listdir(self.input_dir):
-            orig_img = Image.open(self.input_dir + os.path.sep + filename)
-            for basewidth in target_sizes:
-                img = orig_img
-                # calculate target height of the resized image to maintain the aspect ratio
-                wpercent = (basewidth / float(img.size[0]))
-                hsize = int((float(img.size[1]) * float(wpercent)))
-                # perform resizing
-                img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
+        while True:
+            filename = self.img_queue.get()
+            if filename:
+                logging.info("Resizing image {}".format(filename))
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
+                    
+                    # save the resized image to the output dir with a modified file name 
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
+
+                os.remove(self.input_dir + os.path.sep + filename)
+                logging.info("Done resizing image {}".format(filename))
+                self.img_queue.task_done()
+            else:
+                self.img_queue.task_done()
+                break
                 
-                # save the resized image to the output dir with a modified file name 
-                new_filename = os.path.splitext(filename)[0] + \
-                    '_' + str(basewidth) + os.path.splitext(filename)[1]
-                img.save(self.output_dir + os.path.sep + new_filename)
-
-            os.remove(self.input_dir + os.path.sep + filename)
         end = time.perf_counter()
-
         logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
 
     def make_thumbnails(self, img_url_list):
         logging.info("START make_thumbnails")
         start = time.perf_counter()
 
-        self.download_images(img_url_list)
-        self.perform_resizing()
+        map( # Mapping the put method of the queue object to the url list
+            lambda img_url: self.dl_queue.put(img_url), 
+            img_url_list
+            )
+
+        num_dl_threads = 4
+
+        for _ in range(num_dl_threads):
+            thread1 = Thread(target = self.download_images)
+            thread1.start()
+            # Join does not need to be called because a join blocks 
+            # the calling thread until the join thread is complete
+            # The threads are more dependent upon the resizing action
+            # as opposed to the download action
+        thread2 = Thread(target = self.perform_resizing)
+        thread2.start()
+
+        self.dl_queue.join()
+        self.img_queue.put(None)
+
+        thread2.join()
 
         end = time.perf_counter()
         logging.info("END make_thumbnails in {} seconds".format(end - start))
